chore: remove unused code from prev design
