--- frontend/src/components/Navs/homeNav/style.module.css
@@ -1,53 +1,0 @@
-.container {
-  padding: 2vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-  z-index: 3;
-}
-.buttons {
-  background: var(--nav-btn-background);
-  padding: 1vh 2vh;
-  border-radius: 20px;
-}
-.navBtn {
-  margin: 0 2vh;
-  color: white;
-  padding: 1vh;
-  cursor: pointer;
-}
-a {
-  text-decoration: none;
-}
-.navBtn:hover {
-  color: rgb(241, 238, 238);
-}
-.nav {
-  list-style: none;
-  margin-bottom: 5vh;
-}
-
-.vector {
-  width: 100%;
-}
-
-@media only screen and (max-width: 1180px) {
-  .buttons,
-  .nav {
-    display: none;
-  }
-  .container {
-    grid-area: footer;
-  }
-  .vector {
-    width: 80%;
-    display: none;
-  }
-}
-
-@media only screen and (max-width: 490px) {
-  .vector {
-    display: block;
-  }
-}
