--- frontend/src/components/Navs/homeNavMobile/style.module.css
@@ -1,44 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-}
-.container {
-  display: none;
-}
-
-@media only screen and (max-width: 1180px) {
-  .container {
-    display: block;
-    /* position: absolute; */
-    grid-area: nav;
-  }
-  .somesvg {
-    position: absolute;
-    left: 0;
-    margin: 2vh;
-    z-index: 4;
-    cursor: pointer;
-  }
-
-  .navContainer {
-    position: absolute;
-    top: 0;
-    background: #14335f;
-    height: 100vh;
-    width: 100vw;
-    z-index: 6;
-    color: white;
-    list-style: none;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    clip-path: circle(66px at 0 0);
-    -webkit-clip-path: circle(66px at 0 0);
-    transition: all 0.5s ease-out;
-  }
-  .activeNav {
-    clip-path: circle(200vh at 0 0);
-    -webkit-clip-path: circle(200vh at 0 0);
-    position: fixed;
-  }
-}
