--- frontend/src/components/Navs/navLinks/style.module.css
@@ -1,64 +1,0 @@
-.Link li {
-  color: #c7c7c7;
-  font-size: 1.1rem;
-  font-weight: 600;
-  margin: 2vh 0 0 0;
-}
-.Link li:hover {
-  color: #ffffff;
-}
-
-.Link {
-  text-decoration: none;
-  cursor: pointer;
-}
-
-.first {
-  transform: translatex(2vh);
-}
-.sec {
-  transform: translatex(4vh);
-}
-.third {
-  transform: translatex(5vh);
-}
-.fourth {
-  transform: translatex(3vh);
-}
-.fifth {
-  transform: translatex(2vh);
-}
-.sixth {
-  transform: translatex(1vh);
-}
-
-@media only screen and (max-width: 1180px) {
-  .Link li {
-    width: 100vw;
-    padding: 2.5vh 0;
-    text-align: center;
-    margin: 0;
-  }
-
-  .first {
-    transform: translatex(0);
-  }
-  .firstLink {
-    margin-top: 7vh;
-  }
-  .sec {
-    transform: translatex(0);
-  }
-  .third {
-    transform: translatex(0);
-  }
-  .fourth {
-    transform: translatex(0);
-  }
-  .fifth {
-    transform: translatex(0);
-  }
-  .sixth {
-    transform: translatex(0);
-  }
-}
