--- frontend/src/components/shortcuts/style.module.css
@@ -1,42 +1,0 @@
-.container {
-  position: absolute;
-  left: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 10vh;
-  height: 100vh;
-  background: linear-gradient(
-    90deg,
-    rgba(0, 0, 0, 0.17440479609812676) 13%,
-    rgba(0, 212, 255, 0) 100%
-  );
-}
-
-.logos {
-  padding-left: 1vh;
-}
-
-.logo {
-  width: 45px;
-  margin-top: 1.5vh;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-}
-.logo:hover {
-  transform: scale(1.2);
-}
-
-@media only screen and (max-width: 1180px) {
-  .container {
-    right: 0;
-    left: auto;
-    position: absolute;
-    width: 7vh;
-    background: none;
-    z-index: 3;
-  }
-  .logo {
-    width: 5vh;
-  }
-}
